requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
click==8.1.3
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.5
django-debug-toolbar==3.5.0
django-import-export==2.8.0
django-redis==5.2.0
django-rq==2.5.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
redis==4.3.4
rq==1.10.1
sqlparse==0.4.2
tablib==3.2.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
